+++ .gitignore
@@ -285,3 +285,4 @@ gradle-app.setting
 *.hprof
 
 # End of https://www.toptal.com/developers/gitignore/api/java,gradle,intellij,intellij+all,macos,windows
+build
