+++ src/test/java/com/holub/RMIServerTest.java
@@ -1,0 +1,7 @@
+package com.holub;
+
+import static org.junit.Assert.*;
+
+public class RMIServerTest {
+
+}
